requirements.txt
django==2.0.1
djangorestframework==3.7.7
django-rest-swagger==2.2.0
drf-yasg==1.2.2
coverage==4.5.1
model-mommy==1.6.0


